Update label.py
Make the x and y coordinate labels integers from 1 up to and including 64. Whether the agents overlap can now also be labeled by clicking on the image when prompted to. Maximized the screen to better label the images. You have to press on the image to give the x and y coordinates. This was done to prevent Nones.